refacror: create separated folder for secondary functionality
