--- src/checkValidation.ts
@@ -1,8 +1,0 @@
-import { validate } from 'uuid';
-import { BadRequestException } from '@nestjs/common';
-
-export const checkValidation = (id: string) => {
-  if (!validate(id)) {
-    throw new BadRequestException(`This id: "${id}" is not valid`);
-  }
-};
