+++ src/secondaryFuncs/sendRequest.ts
